Move space route setup to manager
